+++ src/components/SubCategory.tsx
@@ -36,7 +36,15 @@ const SubCategory = () => {
     console.log(rolePermission);
 
     // this is subcategorie const is use for perform on status and actions
+    const hasEditPermission = rolePermission.some(
+        (role: { section: string; permission: string[] }) =>
+            role.section === 'subcategory' && role.permission.includes('write')
+    );
 
+    const hasDeletePermission = rolePermission.some(
+        (role: { section: string; permission: string[] }) =>
+            role.section === 'subcategory' && role.permission.includes('delete')
+    );
     const subcetagories_data_col: ColumnProps<Categories>[] = [
         ...SUBCATEGORIES_DATA_COL(currentPage, 10),
         {
@@ -71,57 +79,51 @@ const SubCategory = () => {
                 );
             },
         },
-        {
+    ];
+    if (hasEditPermission || hasDeletePermission) {
+        subcetagories_data_col.push({
             title: 'Action',
             key: 'action',
             align: 'center' as AlignType,
-            render: (_, record: Categories) => {
-                const editPermission = rolePermission.some(
-                    (role: { section: string; permission: string[] }) =>
-                        role.section === 'categories' && role.permission.includes('write')
-                );
-
-                console.log(editPermission);
-                const deletePermission = rolePermission.some(
-                    (role: { section: string; permission: string[] }) =>
-                        role.section === 'categories' && role.permission.includes('delete')
-                );
-                return (
-                    <div className="flex gap-2 justify-center">
+            render: (_, record: Categories) => (
+                <div className="flex gap-2 justify-center">
+                    {hasEditPermission && (
                         <div>
                             <button
                                 onClick={() => handleEdit(record, record.id)}
-                                disabled={!editPermission}
-                                className={`py-3 px-4 rounded ${editPermission ? 'bg-blue-500 text-white' : 'bg-gray-300 text-gray-500'}`}
+                                className="py-3 px-4 bg-blue-500 text-white rounded"
                             >
                                 <FaEdit />
                             </button>
                         </div>
-
+                    )}
+                    {hasDeletePermission && (
                         <div>
                             <button
                                 onClick={() => handleDelete(record.id)}
-                                disabled={!deletePermission}
                                 className="py-3 px-4 bg-red-500 text-white rounded"
                             >
                                 <MdDelete />
                             </button>
                         </div>
-                    </div>
-                );
-            },
-        },
-    ];
+                    )}
+                </div>
+            ),
+        });
+    }
+
     // there is a handle addItem Permissions check
     const addItemPermission = rolePermission.some(
         (role: { section: string; permission: string[] }) =>
-            role.section === 'categories' && role.permission.includes('delete')
+            role.section === 'subcategory' && role.permission.includes('delete')
     );
     //handle status change of subcategories
     const handleEnable = async (id: string, currentStatus: string) => {
         // setStatusId(id);
         try {
             const changeStatus = currentStatus === 'active' ? 1 : 0;
+            console.log('changed status:', changeStatus);
+            console.log('Sending value:', changeStatus);
             const statusUpdate = await http.patch(`/api/v1/subcategories/${id}/?category_id=${params.id}`, {
                 status: changeStatus,
             });
@@ -296,7 +298,10 @@ const SubCategory = () => {
         async (page: number) => {
             // setLoading(true);
             try {
-                const response = await http.get(`api/v1/subcategories/?category_id=${params.id}`);
+                const skip = (page - 1) * 10;
+                const response = await http.get(
+                    `api/v1/subcategories/?category_id=${params.id}&limit=10&offset=${skip}`
+                );
                 // const data = await response.json();
                 // console.log(response);
                 setCategoriesData(response.data.data);
@@ -344,7 +349,7 @@ const SubCategory = () => {
             <div className="flex justify-end mb-4 gap-5">
                 <div className=" ">
                     <Button
-                        type="primary"
+                       
                         onClick={handleBack}
                         style={{ color: '#2967ff', backgroundColor: '#ffffff' }}
                     >
@@ -352,14 +357,11 @@ const SubCategory = () => {
                     </Button>
                 </div>
                 <div className="">
-                    <Button
-                        disabled={!addItemPermission}
-                        type="primary"
-                        onClick={handleAdd}
-                        style={{ color: '#2967ff', backgroundColor: '#ffffff' }}
-                    >
-                        +{ADD_ITEM}
-                    </Button>
+                    {addItemPermission && (
+                        <Button onClick={handleAdd}  style={{ color: '#2967ff', backgroundColor: '#ffffff' }}>
+                            +{ADD_ITEM}
+                        </Button>
+                    )}
                 </div>
             </div>
